Implement parallel training process

+++ src/network/optimization.py
@@ -54,9 +54,10 @@ def accuracy_function(real, pred, strategy=None):
     with context:
         mask = tf.math.logical_not(tf.math.equal(real, 0))
 
-        accuracies = tf.equal(real, tf.cast(tf.argmax(pred, axis=2), dtype=tf.int32))
+        accuracies = tf.equal(real, tf.cast(tf.argmax(pred, axis=2), dtype=tf.int16))
         accuracies = tf.math.logical_and(mask, accuracies)
 
         accuracies = tf.cast(accuracies, dtype=tf.float32)
         mask = tf.cast(mask, dtype=tf.float32)
+
         return tf.reduce_sum(accuracies) / tf.reduce_sum(mask)
